MGS-5226 [#imx-1746] ovxlib-0033-fix-stride-slice-infinite-loop-issue
fix stride slice infinite loop issue

Reviewed-on: http://192.168.32.113/review/5118
Tested-by: kaihong.xia <kaihong.xia@verisilicon.com>

+++ ovxlib/src/ops/vsi_nn_op_strided_slice.c
@@ -190,8 +190,8 @@ static vsi_bool _check_is_same_shape(
     int32_t *stride
     )
 {
-    uint32_t i = 0;
-    uint32_t dims = inputs[0]->attr.dim_num;
+    int32_t i = 0;
+    int32_t dims = (int32_t)inputs[0]->attr.dim_num;
 
     for (i = dims - 1; i >= 0; i --)
     {
